Programming Assignment 2: Lexical Scoping
The goal of this respository is to write two functions that cache the inverse of the matrix. The first function creates a list, and the second function calculates the inverse of the matrix in case it was not calculated (and stores it in the cache), otherwise it will get the inverse of the matrix from the cache.